requirements.txt
asyncify==0.10.0
maafw==2.2.2
nicegui==2.5.0